Switch to using the rosdep liboctomap-dev key.
That is, instead of taking this dependency from the
ROS distribution, take it from the system.  This is
to ensure that we use a consistent ABI with the
platform-provided liboctomap.

Note that in the near future, we'll be removing the
ROS-provided octomap in favor of only using the
system-provided one.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>